refactor: add lib crate and bin cli
