+++ src/lib.rs
@@ -1,0 +1,8 @@
+mod agent;
+mod api;
+mod config;
+mod errors;
+mod tools;
+
+pub use crate::agent::Agent;
+pub use crate::config::ApiConfig;
